payment-channel-sdks/python/pyproject.toml
[project]
name = "near-payment-channel-sdk"
version = "0.1.0"
description = "Near Payment Channel SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "base58>=2.1.1",
    "borsh-construct>=0.1.0",
    "construct-typing>=0.5.6",
    "py-near>=1.1.50",
    "pydantic>=2.10.4",
    "pynacl>=1.5.0",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = ["ipython>=8.31.0", "ruff>=0.8.6"]

[tool.python]
python-version = "3.10"
